Squares
The main square types are now live. The example file only contains step back and reroll but almost all of them are available.
I say almost, 'cause the 'draw a card' square is still under developement(since it needs a deck of cards to work). The next goal is to complete the card developement so that the last square can be completed too.